auth: Keep historical copies of AuthDB entities when modifying them.
This history will later be used to construct a change log (an ordered list of
actions).

I've also considered to generate and store diffs only instead of full copies.
But implementation looked no less hairy (there was lot of diffing code in
various places that update groups). Also if some detail is forgotten in the code
it will be gone forever.

With approach proposed in this CL all information is preserved in full, and
change history can be rebuild any time. Also all diffing code can be
concentrated in a single place (will be done in next CL).

R=nodir@chromium.org
BUG=luci/luci-py#216

Review URL: https://codereview.chromium.org/1263143010 .