Make deleted bot visible through web UI as long as there were events.
This permits investigating failure with autoscaled bots; the autoscaled bot is
not visible in the bots list but if a user wants to investigate what happened in
the past on a bot that is since deleted, it can now be done.

While at it, make the page faster by using more asynchronous NDB operations.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1303573002 .