Trap NeedIndexError in cron job cron_handle_bot_died.
This is annoying when deploying a new instance, as the cron job start right
away but the indexes takes several minutes (5~6 from observation) to be created
even if there is no entity to index yet. This is very confusing for the user
deploying the code and give no value. Silence it.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1313663002