Improve handling of autoscaler shutting down a VM in the middle of a task
In practice we'd prefer this to never happen. Until then, report something more
meaningful to the user and most importantly do not register an exception to
reduce the noise in the server logs.

Most importantly, define a clear way for task_runner to report an "expected"
internal error (vs just a crash) with text based on event. In that case, it's
receiving a signal.

Create a complete smoke test to ensure signal handling is correctly handled and
works on Windows too.

Factor out FakeSwarmingServer into its own file.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1324163002