Make all swarming bot subprocesses detached.
This is required on Windows for proper signal handling at process granularity.
Without this, signals are sent to 'process group', which is all the processes
created that are not detached.

detached=True on POSIX will also cause the child processes to be in their own
process group via os.setpgid(0, 0) but this doesn't cause any issue.

- Standardize on subprocess42 in all relevant places.
- Fix a bug in subprocess42 in case of exception, where the signal handlers
  would not be restored.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1322003003