Fix regression in dc1706c7374b02.
sys.platform=='linux3' is useless as a signal. The AppEngine SDK leaks something
more useful; os.__file__ is still set and will not start with '/' when run under
Windows. Use that as a quick and reliable signal to know if a unit test or dev
server is running on Windows.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1319023005