Small tweaks to the endpoints APIs; make more APIs to be GET.
In practice someone would be running around yelling but it's fine because this
API is not used yet. This makes all non mutating requests as GET, as it was a
bit inconsistent.

Convert DateTimeField to FloatField in these GET requests, as there's a bug in
endpoints-protorpc integration when passing a DateTimeField as a query argument.
It simply doesn't work.

Change start and end to default to None (no boundary) instead of respectively 24
hours ago and start + 24 hours.

R=vadimsh@chromium.org
BUG=https://github.com/luci/luci-py/issues/197

Review URL: https://codereview.chromium.org/1308893007