Remove 'work' on bot start.
'work' may be present from a hard reboot or other failure mode. The fact that it
may be full of junk may cause the bot to auto-quarantine itself, leading to a
deadlock, the bot won't delete 'work' until it's given a task, but it won't be
given a task because it doesn't have enough free space.

This will unquarantine ~65 bots on prod.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1358463002