Fix error when task_update fails (HTTP 500) on the closing request.
This caused bot_update_task() to be confused and not mark the task as completed
due to setting 2 exit_codes by error.

This eventually lead to the task being left as if it hadn't completed, causing
an internal_failure when the BOT_DIED reaping cron job noticed lack of activity.

Rename .duration to .duration_total to clarify that this includes overhead. This
is done in preparation of migration .exit_codes and .durations respectively to
.exit_code and .duration. In this case, .duration won't include overhead. Reduce
the use of .exit_codes.

R=vadimsh@chromium.org
BUG=chromium:538744

Review URL: https://codereview.chromium.org/1403933004