Fix bug and improve logging in case of failure.
This was causing BOT_DIED tasks, but this is just one symptom.

There's something else wrong in the code that causes TaskRunResult.started_ts to
not be set, which is in theory impossible. I'll have to continue to investigate
but this will help a bit.

R=vadimsh@chromium.org
BUG=chromiumr:538744

Review URL: https://codereview.chromium.org/1401293003