Allow backends to specify attributes to include in Pub/Sub messages
Arbitrary attributes for any purpose. Machine Provider will include them verbatim.

Current use:
GCE Backend tells the Machine Provider to include an attribute "group" whose value is the instance group name that the instance belongs to. This is so that when the GCE Backend gets a reclamation request, it can actually figure out where the instance came from (since the instance name alone is not the key to any entity).

Possible future use:
Could also be used to add a secret that Machine Provider must transmit along with the messages. Can be used with push subscriptions to ensure the message being processed is really coming from the Machine Provider.

BUG=489837

Review URL: https://codereview.chromium.org/1407183004