WaitUntilFullyBooted: wait for the package manager service to be up and running.
This is necessary as many unit tests expect the device to be in working
condition when they start which includes the device to be booted up to a point
where Package Manager is running.

The error message
  'Error: Could not access the Package Manager.  Is the system running?'
could be observed from the tests.

Add an assert for now to ensure the output is what is expected. Will relax it if
needed.

R=stip@chromium.org
BUG=chromium:525873

Review URL: https://codereview.chromium.org/1416373009