Soften the auto-quarantine for low disk space.
Account for small partitions which may be small (e.g. 10Gb on ChromeOS). In this
case it is not realistic to ever have 4Gb free. Add a second threshold, which is
the fraction of free disk space that must be free. Default to 15% to for a 10Gb
partition, this means the effective threshold is lowered from 4Gb to 1.5Gb.

R=vadimsh@chromium.org
BUG=chromium:545672

Review URL: https://codereview.chromium.org/1419053005