Fix tags being overwritten when a task is deduplicated.
Only the copy in TaskResultSummary was corrupted, not the one in TaskRequest. So
this was only a problem when searching for TaskResultSummary by tags, the UI
wasn't affected.

Add mapreduce job to clear up the corruption and also backfill previous
entities, which will be run a few hours after switching the default version.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1455983003