Add test to clarify which requests are supported and which aren't.
Better handle the situation when a request is unsupported. Returns a meaningful
error instead of a HTTP503.

I plan to remove stale indexes on prod, so I want to make sure the query options
product work as expected. The end goal is to limit as much as possible composite
indexes on TaskResultsummary, are they incur an enormous overhead on very hot
entities.

Remove an unneeded import db.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1472763005