Remove stale reference to old /content-gs/ isolate API.
Update /browse to not do XHR request anymore as it was using the old API. It
will be switched to use the Endpoints API once the web UI is switched to OAuth2
based authentication. Until then, it's impossible to use the new API as the Web
UI login is using AppEngine's authentication, which is not understood by
Endpoints.

In the meantime the resulting code is a bit cheezy but THIS IS TOTALLY TEMPORARY
CODE, right? It's fairly possible the server is rewritten in Go before this
happens.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1523183002