Allow users to give machines instructions to execute as soon as they're leased
In general the user will be able to give instructions to a leased machine at any time, but also allow them to give an instruction to execute immediately on lease. For now this is limited to a Swarming server (i.e. on lease fulfillment connect to a Swarming server immediately).

BUG=489837

Review URL: https://codereview.chromium.org/1528203002