Zap out more spurious error report.
This happens when the bot reboots but the original swarming_bot.zip was running
an older version. In the Windows case, the parent process is still the old
version of the bot, which is why there's this error report in the first place.

Just ignore it.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1619423004