Add audio state on bots.
It doesn't add it as a dimension yet to first verify it works well. On OSX it
tries to guess if the audio cable is connected but this wasn't tested (I don't
have a mac mini to confirm).

Tested for it to work on each platform manually.

R=tandrii@chromium.org
BUG=chromium:574174

Review URL: https://codereview.chromium.org/1639743003