Fix a test case due to ab5f4bc.
This the test hacks out NDB functions, the fix in ab5f4bc changing which
function is called as part of the transaction changed the outcome of this test.
I still prefer to keep the test since it ensures that it somewhat behaves in
case of aggressive DB failures.

R=nodir@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1770853002