Set queue tag and verify to 500/second & max_concurrent_requests=10000.
The maximum rate of 500/s has been supported since GAE 1.5.0 as noted in the
RELEASE_NOTES file but is not otherwise documented elsewhere.

The maximum number of concurrent requests has been made configurable since GAE
1.4.2 and defaults to 1000.

There is currently 460k accumulated 'tag' task queue items. Hopefully with the
new settings we can expect the queue to empty in ~1h but we'll have to keep an
eye on this as 500/s is the maximum supported rate. :/

R=sergeyberezin@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1776163002