Remove TaskRequest.data. This hasn't been supported for several months.
The only way to use it was through the old API and this API was deleted already.

This means that older entities will have their data URL lost. Since it was on
isolate servers that we know the items have expired, it's a non-issue now since
the data is already irrecoverable.

A side effect of this change is that task property hash changes, which means the
task result can't be resused when crossing this commit. It's not a big deal.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1800243002