swarming: change meaning of inputs_ref
Currently inputs_ref implicitly specifies isolated server and namespace
for isolated outputs. The problem is that a user may need to be able to specify
them for non-isolated command.

Change the meaning of inputs_ref from "a reference to input file tree"
to "isolate server/namespace and hash of input file tree".
Now (inputs_ref is not None) does not mean there is isolated input,
but (inputs_ref and inputs_ref.isolated) does.

In bot manifest, change "inputs_ref" to "isolated" that contains
input/output settings.

Add a TODO to refactor the code and update datastore schema.

Update tests.

R=maruel@chromium.org
BUG=601022

Review-Url: https://codereview.chromium.org/1939343002