swarming: refactor cipd input
Add cipd server settings that specify cipd server host and cipd client package.

Refactor "packages" field in task properties and manifest to "cipd_input"
message that contains the packages and cipd settings.

Allow parameters in CipdPackage.package_name. A bot will provide their values.

This CL changes TaskProperties schema in a backward incompatible way, but we never used "packages" datastore
field so far, so it is not a problem.

R=maruel@chromium.org
BUG=601022

Review-Url: https://codereview.chromium.org/1946253003