Use a single cron job to enqueue all metadata-related tasks
Since there is a precedence order which causes certain tasks to abort if others are in progress, no point in scheduling ones that will be aborted anyways.

Also makes InstanceGroupManagers.instances refer only to instances returned by the GCE API, and adds additional logging relating to detecting and deleting instances.

BUG=609308

Review-Url: https://codereview.chromium.org/2050523002