Update some Polymer 0.5 stats to be Polymer 1.0
The Polymer 0.5 code has been put in its own folder to separate it from the new Polymer 1.x elements.  All but the stats-app element have been converted.  stats-app will use some of the new selectors I'm working on, so I've decided not to convert it quite yet.

Demo of 1.x code:
https://screenshot.googleplex.com/KLeTVUnnWch

This also adjusts the places that used the Polymer 0.5 stuff to be in their own place.

Polymer 0.5 won't continue living for much longer, as it will soon(tm) be migrated

BUG=

Review-Url: https://codereview.chromium.org/1948873003