Implement preliminary support for live task cancelation.
This is useful when a task has a long hard timeout and hangs or too many long
tasks were triggered by accident or in whatever situation where manual load
shedding is needed.

This change doesn't actually implement RUNNING task cancelation on the server
side, it implements the bot side logic to accept a message from the server to
kill the currently running task_runner instance.

R=tandrii@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2027763002