'Fix' flakiness in task_runner_test.py.
The failing test cases are broken by design. The fix would be to have them loop
with increasingly high timeout. For now, increase the timeout, add logs, cry a
little.

Most importantly, it prints out run_isolated logs from task_runner_test, which
is useful to figure out the root cause. In this case, it was that run_isolated
took more than 300ms (!) to startup, so the child program never had the time
to start. I think it's more a testament of the sadness of python than anything
else. Still, these test cases is somewhat broken by design due to the hard coded
timeouts.

R=vadimsh@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2071353002