run_isolated: change cipd package flags
Replace --cipd-package-list flag with --cipd-package that can be
specified multiple times.

Current implementation accepts a path to file with packages instead of a
list of packages to avoid hitting command line limits on Windows.
However it is better to guard from that at the global run_isolated.py
level, that is accept a file with all flags, not only cipd packages.
The mistake was discovered in
https://codereview.chromium.org/2085253002/

R=maruel@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2097443004