endpoints -> webapp2 adapter
Add a function that converts a Cloud Endpoints service class to
a list of webapp2 routes.

Add a base class for tests. The test code for classic cloud endpoints
can be left intact, except call_should_fail function is properly
deprecated.

Had to change auto_stub so it is usable in infra.git

R=vadimsh@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2117833003