Make OAuth client id accessible to new ui w/o hardcoding
The (small) index.html for each ui page will be turned
into a template with one param - client_id

My first try was to make an API endpoint, but that was
very very slow, because we had to wait for the api to
return (1-3 seconds) before making any other requests.
This is zippy and could be memcached, if needed.

BUG=631047

Review-Url: https://codereview.chromium.org/2242543002