Fixup __init__ files in luci's vendored protobuf
It turns out that google.protobuf changed the way it exposes itself in the
'google' package. The changes *probably* make it a lot easier to do things
like use both google.protobuf and google.appengine side-by-side. But we
have our own mechanisms in place to make sure that works, and whatever these
changes do they don't play well with our mechanisms. So this CL resets the
__init__.py files to set up the package the same way as it used to.

R=sergeyberezin@chromium.org
BUG=monorail:1619

Review-Url: https://codereview.chromium.org/2255143003