Add filter to task-list
The two big things in this CL are adding in the filtering mechanism (very similar to filtering bot-list) and migrating task-list-data to use sk.request.

In the task list, you can dynamically choose your columns and quickly see the state of tasks, and cancel pending tasks.

Assembling the tags (in task-list-data) is a bit trickier than the dimensions was for the botlist because there are many more tasks than bots, so the cron job can only run once per hour for the last hour's worth of data.

BUG=631047

Review-Url: https://codereview.chromium.org/2266133002