Import _strptime very early in bot_main lifetime.
This works around a race condition in python library. The hack was done in
Chrome's internal bot_config.py but it's better to just do it unconditionally.
Technically there was probably a reason why this package is lazily loaded but
the rationale was lost in the abyss of python standard library history. Since
we're already doing it anyway in bot_config.py, let's just do it for everyone.
This will remove dead code from our bot_config.py.

R=vadimsh@chromium.org
BUG=584730

Review-Url: https://codereview.chromium.org/2286323003