Introduce new bot-page UI
The bot page is relatively simple (compared to the dynamic tables and
query-able filters used by bot-list and task-list)

It shows the bot's dimensions and other pertinent data.  There are buttons
for doing some bot maintenance (e.g. deleting dead bots), but those don't
quite work yet - next CL should add it.

In the demo code, you can view test-bot-001, 002 and 003 to see a normal
bot, a quarantined bot and a dead bot respectively.

The state is currently just a pretty printed version of the state object
that comes with the bot.  Perhaps in the future, important data from there
(like devices) can be pulled out and shown more easily.

BUG=631047

Review-Url: https://codereview.chromium.org/2291323002