Adds auth option `luci_context_json`.
Adds the CLI option for `luci_context_json` as well as support for
LUCI_CONTEXT environment variable. The LUCI context is a JSON-encoded
set of parameters, one of which is `local_auth`. This parameter can be
used for querying a local server for an access token. The local server
will handle retrieval of a valid token from an external source, and is
used to mask this interaction from a child (unpriviliged) process
running a task.

This also adds a dummy class, LuciContextAuthenticator, which will
ultimately handle the interaction with the local server and distribution
of the access token to unpriviliged tasks.

R=vadimsh@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2318993003