Refactor prior to adding task-page
There should be no real visual changes here. This is minor bug fixes and
refactoring to aid in the addition of task-page.

This fixes the name of "limit"->"_limit" in the two filters.  The latter
was declared but the former was used.

_taskLink and _botLink are consolidated into the common behavior to avoid
duplication (and fixing the links in multiple places)

Many styles that will be used by task-page are pulled out of bot-page and into
single-page-style.

_timeDiff* are pulled into common behavior (they will be used in task-page)

_humanDuration is created in common behavior (to consolidate and allow for use
in the templates of task-page)

buttons now look more clickable (pointer cursor)

task-list sorts times better when they don't have one

task-list uses human.localeTime like the rest of the app

The pages titles have been changed to be more useful when more tabs are open.
Instead of "Swarming Bot List", it is now "Bot List - Swarming", which means
if you have 20 tabs open, you see "Bot Li" instead of "Swarmin", which makes
it easier to find the right tab.

BUG=631047

Review-Url: https://codereview.chromium.org/2338383002