Make the buttons on task-page work
You can now cancel and retry tasks from the task-page.

When retrying a task, all the current properties of the task are
copied into the new task and "(retry)" is appended to the name.
Additionally, the user field is set to be whoever is logged in.
If the retry is successfully requested, the task page will auto
redirect to it.

The refresh interval actually works, making the stdout and milo refresh.

The Requested Dimensions now has a link to the botlist with
all those dimensions, so one can see which bots could pick up a
given task.

The times on task-page now have their full timestamps as mouseovers.

BUG=631047

Review-Url: https://codereview.chromium.org/2350853004