Add pageable data widget
This allows the user to request more data if they need it.
This has been integrated into bot-page and can be added
to bot-list and task-list as well.

So far this seems to speed up the bot-page a bit because
the initial query for data is smaller.  It seems that the
previous logic of paging for the last 200 tasks was the most
expensive part of the whole page.

BUG=631047

Review-Url: https://codereview.chromium.org/2372323002