auth: Overhaul delegation.proto messages.
This is done in preparation for adding support for service account tokens.

Changes:
  * Switch delegation.proto to proto3 syntax.
  * Rename 'issuer_id' to 'delegated_identity'. It makes more sense.
  * Rename 'impersonator_id' to 'requestor_identity', make it always present.
  * Add new 'kind' field (that will be different for service account tokens).

The wire format and all REST APIs are unchanged.

Also update 'compile_protoc.py' script to always use protoc in PATH and require
it to be 3.0.0. protoc 3 is available as part of infra.git environment (or can
be installed in any other way).

R=maruel@chromium.org, nodir@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2379213002