Fix background page loading
You should now be able to open pages in a new tab
and they will start loading.

I was able to get rid of the popup that would happen
sometimes by being a bit more patient on the auth
library to load its thing.

The main culprit was the fact that we were waiting until
the auth-signin's attached phase to do anything, which
apparently requires tab focus.  The ready phase can fire
while the tab is still hidden, so we use that instead.

I also made the token "offline", seems to have helped with
more consistent credential management by the
google-signin-aware element.  It may have been unnecessary,
but it works nonetheless.

BUG=653703

Review-Url: https://codereview.chromium.org/2401763003