LRUDict: add timestamps
lru.LRUDict currently stores the order of items, but not time when an
item was used most recently.

Add timestamps to LRUDict internal representation and update persistence
format. Preserve ability to load a dict from the old format for backward
compatibility.

Also remove lru.LRUDict.batch_insert_oldest because it complicates this
CL. It wasn't used.

R=maruel@chromium.org
BUG=654562

Review-Url: https://codereview.chromium.org/2409103003