luci-py/isolateserver.py: Add support for tar archives when downloading.
Currently the isolateserver.py only supports extracting ar archive which
contain many small files grouped together. This CL extends the support
to also include tar archive format.

This support will exist while we evaluate and compare the ar and tar
file formats with real workloads on real servers.

It is expected that one format will be removed in the future.

 * https://github.com/luci/luci-go/issues/9
 * https://crbug.com/598990

BUG=598990

Review-Url: https://codereview.chromium.org/2484133002