Make silent task retries more obvious
This adds a task-disambiguation element that shows the
(silent) retries, their state and the bot.

Since milo-pages are not generated for tasks not ending
in 0, I added an explanation for when this happens.

This also tweaks the gpu sorting order on the botlist
page.

I had considered making the task list on the bot-page link
to the non-summary version of the task, but I decided
the task-disambiguation explains more (i.e. the task was
retried).

BUG=661152

Review-Url: https://codereview.chromium.org/2497663003