Move error-toast up and Give a warning if clientid is not set
This makes sure all pages have an error-toast element. This allows
server details problems (typically 403) easier to see and address,
especially when setting up a new instance.

The pages now will be bright red and easily visible if client id
is not set.  Again, this is useful for setting up a new instance.

BUG=665442

Review-Url: https://codereview.chromium.org/2511463003