Add informational messages to clarify what can be filtered on.
These informational messages should clarify that users can only filter
on tags and dimensions.

This adjusts the "float selected columns to the top" to sort those
floated columns in the order they are displayed.

This adds ... at the bottom of the task list to clarify that
not all options are givin for sorting tags (due to memory constraints).

This changes sort order a bit.  "none" now sorts at the bottom (like
ZZZ) which makes it less obnoxious than being in the middle.

This adds internal_ip to visually see what ip addresses have been
assigned.

BUG=665277,664832

Review-Url: https://codereview.chromium.org/2513613002