Add task-list summary
This cleans up bot-list-summary as well, given the two
elements are related.

This removes the restriction server side that the end time
must be given for a count - as that just makes queries a
bit more awkward to construct. The start time is definitely
required, as counting things from the beginning of time would
time out.

BUG=673774

Review-Url: https://codereview.chromium.org/2676513004