swarming_bot: Rename 'restart' to 'host_reboot'
Soon a 'bot_restart' will be added to be able to restart the bot process without
rebooting the host. The rename, including adding a prefix, is to help clarify
what is what.

Include compatibility symbols until all the bot_config.py are updated.

Splitting the rename from the enhancement (in a follow up) to simplify
reviewability. No functional change, except supporting the new command
'host_reboot'.

R=vadimsh@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2691103003