swarming_bot: refactor task_runner.run_command()
Create object RunState to store and handle state changes and handle RPCs.
Increase the amount of comments to help unfamiliar developers understand more
the control flow. Makes the loop much more comprehensible (I hope).

"Mostly" non-functional change. Only one test expectation was updated because it
the previous code was incorrect but this CL tries to make not visible change as
much as possible.

In practice an "exit_code cleanup" should probably be redone later but this CL
is already complex enough.

R=vadimsh@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2706293012