Remove log for missing whitelist in acl check.
Currently, the call to is_ip_whitelisted_machine() which is used to
verify if the machine is IP whitelisted ouptuts a warning if the IP
whitelist groups cannot be found (including the default <appid>-bots
group). In this scenario, the warning log is not required, so remove it.

The previous implementation was causing warnings to be logged when IP
whitelisting was not used as a form of bot authentication.

R=maruel@chromium.org, vadimsh@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2732173002