swarming: send pubsub notifcation on all state transitions
Swarmbucket does not receieve a notification when a task starts, only
when it reaches a terminate state. Because of that swarmbucket has to mark a new
build as started, so it is displayed yellow in code review. This is a
lie. It is confusing to click on a yellow build bubble and see a PENDING
build on Milo.

Change swarming pubsub notifications to be sent on each
TaskResultSummary.state transition, not necessarily to a terminal state.

R=maruel@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2744613003