swarming: Migrate internal use from SHA1 to SHA256.
Migrate both:
- TaskProperties.properties_hash
- swarming_bot.zip hash

python 2.7 doesn't expose SHA3 or BLAKE2 so SHA256 is the next less bad choice.
There's going to be a performance hit but it should be reasonable since most
hashes are cached.

Try to consistency use SHA256 or sha256, so the next algorithm replacement is
easier.

Update a documentation error in the detailed design doc stating that TaskToRun
uses SHA1. That's not true, it uses MD5. :) In practice it doesn't matter since
this is only used as a coarse first pass filter.

Update many string constants in task_request_test.py with the u prefix to make
the diff easier to read, for the next time we update the hashing algorithm.

This will cause all the bots to self-update (like usual).

The only user visible change is that previous task will not be deduped until a
task with the SHA256 component is triggered.

R=vadimsh@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2710213006