Instruct machines to connect to the Swarming server after creating BotInfo
Rather than instructing them to connect as soon as the lease is fulfilled.

This fixes a race condition where the machine connects to Swarming so fast that the BotInfo hasn't been created with MP-related fields yet, so it gets placed in the default pool in bots.cfg even if it really belongs in another pool. Also fixes a race condition where we were trying to create the BotInfo with MP-related fields at the same time the bot handshake handler was creating the BotInfo without MP-related fields, which was overwriting the MP-related fields with None. Now BotInfo with MP-related fields will exist before an MP bot connects to Swarming.

BUG=681224,699797

Review-Url: https://codereview.chromium.org/2767353002