config: store binary configs
config.get_async: if last good config is requested and destination type
is a protobuf message, store it in binary in addition to raw format.

When loading from a last good conig, if protobuf message type name
matches what we have in the datastore, load the config from the binary
field.

Also fix tests and include them in PRESUBMIT.py

R=vadimsh@chromium.org
BUG=705002

Review-Url: https://codereview.chromium.org/2778533002