Fix switching version from an old branch
Prior to this fix, if a user were on an old branch, but wanted to roll
the default version forward, tools/gae would err out with something like
b929dd8854b52b2abb08~-3..b929dd8854b52b~-15 is ambiguous because the revision
was based on HEAD (where the user is).  Changing the flag to --independent
simply gets the revision of the remote branch (e.g. origin/master) and does
calculations relative to that - which is the intended behavior.

BUG=

Review-Url: https://codereview.chromium.org/2857703002