isolate: make /content auto-delete the DB entity when missing file from GCS
Under very rare circonstances, a file may be in the DB but not in GCS. For
example someone called the finalize_gs_upload without successfully uploading the
file, or someone deletes the file from GCS later, or an internal bug.

In this case, /preupload always tell the file is there even if it is not.
Browsing for content returned an HTTP 500.

Instead of HTTP 500, delete the DB entity transparently.

Improve testing a bit.

Fix handler to return HTTP 400 when digest is missing.

R=iannucci@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2894463002