Fix TaskDimensions refresh.
The algorithm was incorrect, causing TaskDimensions.valid_until_ts to be capped
at 1h (_ADVANCE) instead of expiration_ts + _ADVANCE. This caused the
TaskDimensions instances to be evicted too early.

The observed sympthom is specifically for tasks dimensions that where triggered
in a burst but created more that 1 hour of backlog. Eventually the backlog
failed to be processed once tidy_stale() got executed and evicted the
TaskDimensions too quickly.

Update all valid_until_ts checks to be inclusive. Generally end time is
exclusive but I felt it right to be inclusive here.

Update unit test to be clearer about testing of expiration.

Change the task queue to use json encoding, to pass more argument.

Small change to utils.py.

BUG=704236

Review-Url: https://codereview.chromium.org/2903463003