Revert "task_queues: Switch the scheduler to use the new multi-queues algorithm"
This reverts commit c5de71bd7d2ea78ce79047121b666098c7264f87. There was no merge
conflict.

The scheduler is not broken. It's just that when the scheduler is enabled, it is
preferable for 2696794759139f7fcf96 to be live, otherwise pending tasks like CT
and perf will not be correctly scheduled. This CL is to avoid this problem
specifically with these tasks: burst of >1h worth of tasks, then no trigger,
leading to TaskDimensions to incorrectly expire.

So the plan is:
- revert the new scheduler (this CL)
- deploy this new version everywhere, which is essentially 83275bf1f61bac16 with
  2696794759139f7fc cherry picked
- wait 8 hours
- revert this CL
- deploy again, now with the new scheduler

TBR=vadimsh@chromium.org
BUG=704236

Review-Url: https://codereview.chromium.org/2901873002