Improve Schemas.md and in-source schemas
No code change. Only beautiful ASCII Graphs! The best CL!

This is useful to have a better picture of all the pieces.

Standardize them to be all similar and use the same spacing, for coherency.
The Swarming entity schema is complex enough to warrant its own ASCII graph
style guide!

- +4 identation (so it's compatible with markdown)
- Outgoing line down from an entity is aligned at +4 from the entity box's left
- Incoming vertical line goes down at +4 from the entity box's left
- Key ids that are variables are within < and >
- Entity boxes are separated with two spaces, unless it's a series, then it's
  separated with " ... " both at the line and twice between the boxes.
- Embedded entities are a box inside another entity box, with no id line.

R=vadimsh@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2897353002