Run rebuild_task_cache() inline for 'id' task.
This permit quicker execution of these tasks as they can only run on a single
bot anyway. A good example is 'terminate' tasks that are rarely run (once for
each MP GCE bots), so the task effectively has to wait for the GAE task queue to
run to enable the scheduling of this swarming task.

Add unit test to confirm the behavior.

R=vadimsh@chromium.org
BUG=704236

Review-Url: https://codereview.chromium.org/2904953003