Add more input validation.
Refuse duplicate keys in dimensions or env, which means that the user
misconfigured their client.

Specify limits for the user specified inputs. The limits are lenient as this is
really just to detect incorrect inputs or obvious abuse but there is no
significant constraint except for Cloud DB entity size.

R=vadimsh@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2910043002