Add an assert to verify TaskToRun filters.
In issue 727953 a task was not scheduled properly even if the entity was saved
correctly, and all the logs are there to report that the task should have been
scheduled properly. The only thing I can think of is that low_part overflowed,
so adding an assert that this hypothesis is incorrect.

R=vadimsh@chromium.org
BUG=727953

Review-Url: https://codereview.chromium.org/2919593003