Add exception logging on HTTP 400s.
Added to auth/endpoints_support.py even if it doesn't belong there, as this
'just works' for all APIs. I feel that even if it is a layering error and that
it should have been a separate middleware, it's not worth the trouble.

By default endpoints doesn't log anything on HTTP 400 but the error message (and
its stack) are often very important to debug what happened.

R=vadimsh@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2933483006