Promote 'locale' from state to dimension.
This opens the door to scalable non-US English testing.

There are bots which locale detection does not work. In this case, there is no
value set. Ultimately, we'd configure all test execution to specify the desired
locale and a CI could run occasionally the tests in a non-English locale to
confirm the tests are not GUI locale dependent.

Exposing this as a dimension is the first building block.

R=smut@chromium.org
BUG=567838

Review-Url: https://codereview.chromium.org/2932113002