Tweak aggressivity of the poller.
The TaskToRun entities are queried on their 'queue_number' property. When it is
not None, it means the task is ready to be reaped by a bot and run.

As thousands of bots concurrently query for work, a lot of contention exists on
the first TaskToRun entitiy to be returned. The negative cache exists as a quick
and light mechanism to reduce contention on the DB that is already very hot.

This is done by saving in memcache the task IDs that are about to be reaped, so
a concurrent /poll handler knows it is better to skip to the next TaskToRun in
line.

This change tweaks the negative cache to be set before the transaction instead
of after, to make it much more effective.

- Set the negative cache *before* reaping, expiring or canceling a task, to
  reduce NDB transaction contention. It used to be set after successful reaping,
  which means the TaskToRun entity was still hit hard my multiple bots
  concurrently.
- Reduce the negative cache TTL from 2 min to 15s, to reduce the effect of
  transaction failure in PENDING time latency.
- Document more the rationale of the current behavior.

R=smut@google.com
BUG=

Review-Url: https://codereview.chromium.org/2933083004