Split admin.globalConfig into two endpoints
Currently admin.globalConfig claims it can read and write the GlobalConfig,
but if you send an empty GlobalConfigMessage in the request it will actually
overwrite the GlobalConfig with an empty config. In other words, it's not
possible to use the current endpoint to read the config without also writing
it.

This creates separate admin.readGlobalConfig and admin.writeGlobalConfig
endpoints, and makes it impossible for admin.readGlobalConfig to overwrite
the GlobalConfig.

BUG=733843

Review-Url: https://codereview.chromium.org/2941353003